refactor: move deck showcase into component
