+++ src/pages/landing-page/deck-showcase/deck-showcase.scss
@@ -1,0 +1,53 @@
+@import '@/assets/styles';
+@import '@/assets/colors';
+
+.showcase-deck-preview {
+  display: flex;
+  gap: 16px;
+
+  position: absolute;
+  top: -100px;
+  right: -75px;
+  transform: rotate(20deg);
+  overflow: hidden;
+
+  // remove constraints on small viewports
+  @media screen and (max-width: $large-screen) {
+    position: initial;
+    transform: none;
+    overflow: visible;
+  }
+
+  // showcase columns should be hidden on small viewports
+  .column {
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+
+    @media screen and (max-width: $large-screen) {
+      display: none;
+    }
+  }
+
+  .deck-showcase-cover {
+    color: $black;
+
+    .c-front {
+      @include glow-box-shadow($spread-radius: -4px);
+      border-radius: 16px;
+    }
+  }
+}
+
+@keyframes showcase-item {
+  0% {
+    opacity: 0;
+  }
+  50% {
+    opacity: 1;
+  }
+  100% {
+    opacity: 1;
+    transform: scale(1);
+  }
+}
